requirements.txt
sanic
jinja2
pycryptodome
requests
aiohttp